reuse commadimage across protocol editor and run view
