+++ controller/components/runs/list/SwimLaneCommandComponent.tsx
@@ -10,7 +10,6 @@ import {
   MenuList,
   MenuItem,
   IconButton,
-  Image,
   useColorModeValue,
   Center,
 } from "@chakra-ui/react";
@@ -18,16 +17,10 @@ import { HamburgerIcon } from "@chakra-ui/icons";
 import { IoPlaySkipForward } from "react-icons/io5";
 import { BsSkipForwardFill } from "react-icons/bs";
 import { VscRunBelow } from "react-icons/vsc";
-import { PiToolbox } from "react-icons/pi";
 import { useRef } from "react";
 import { capitalizeFirst } from "@/utils/parser";
-import { FaRegFileCode } from "react-icons/fa6";
 import { RunCommand } from "@/types";
-import { TbMessageReport } from "react-icons/tb";
-import { MdOutlinePauseCircleOutline } from "react-icons/md";
-import { MdAlarm, MdStop } from "react-icons/md";
-import { FaRegStickyNote } from "react-icons/fa";
-import { GoStop } from "react-icons/go";
+import CommandImage from "@/components/tools/CommandImage";
 
 interface LaneCommandComponentProps {
   command: RunCommand;
@@ -50,54 +43,6 @@ const SwimLaneCommandComponent: React.FC<LaneCommandComponentProps> = (props) =>
   useEffect(() => {
     toolNameRef.current = toolName;
   }, [toolName]);
-
-  const ToolBoxCommandIconMap = (commandName: string) => {
-    const commandIconMap = {
-      run_python_script: <FaRegFileCode style={{ width: "100%", height: "50px" }} />,
-      show_message: <TbMessageReport style={{ width: "100%", height: "50px" }} />,
-      pause: <MdOutlinePauseCircleOutline style={{ width: "100%", height: "50px" }} />,
-      timer: <MdAlarm style={{ width: "100%", height: "50px" }} />,
-      note: <FaRegStickyNote style={{ width: "100%", height: "50px" }} />,
-      stop_run: <GoStop style={{ width: "100%", height: "50px", color: "red" }} />,
-    } as Record<string, JSX.Element>;
-
-    return commandIconMap[commandName] || <PiToolbox style={{ width: "100%", height: "65px" }} />;
-  };
-
-  function renderToolImage(config: any) {
-    if (!config) return;
-    if (!config.image_url) {
-      return <Box></Box>;
-    } else if (config.name == "Tool Box") {
-      return (
-        <Box display="flex" justifyContent="center" alignItems="center" py={3}>
-          <IconButton
-            aria-label="Tool Box"
-            icon={ToolBoxCommandIconMap(command.commandInfo.command)} // Ensure the icon fills the button
-            variant="ghost"
-            colorScheme="teal"
-            isRound
-            onClick={() => onCommandClick(command)}
-          />
-        </Box>
-      );
-    } else {
-      return (
-        <Box display="flex" justifyContent="center" alignItems="center">
-          <Image
-            src={config.image_url}
-            alt={config.name}
-            objectFit="contain"
-            height="65px"
-            width="65px"
-            transition="all 0.3s ease-in-out"
-            cursor="pointer"
-            onClick={() => onCommandClick(command)}
-          />
-        </Box>
-      );
-    }
-  }
 
   const runningBg = useColorModeValue("teal.200", "teal.800");
 
@@ -179,7 +124,11 @@ const SwimLaneCommandComponent: React.FC<LaneCommandComponentProps> = (props) =>
         </Box>
         <Center p={0}>
           <VStack spacing={2}>
-            <Box>{renderToolImage(infoQuery.data)}</Box>
+            <CommandImage
+              config={infoQuery.data}
+              command={command}
+              onCommandClick={onCommandClick}
+            />
             <Box bottom={0} position="sticky">
               <Text>{capitalizeFirst(command.commandInfo.command.replaceAll("_", " "))}</Text>
             </Box>
